리팩토링 BoardModel -> BoardView2 (추후 이름 수정)
