+++ OrderMate/ViewModel/Setting.swift
@@ -1,10 +1,3 @@
-//
-//  Setting.swift
-//  OrderMate
-//
-//  Created by yook on 2023/03/06.
-//
-
 import SwiftUI
 
 var listItem = ListItem()
